Cambiados los tipos de vars para ahorrar mem

+++ src/ejercicios/Ejercicio3.java
@@ -4,7 +4,7 @@ public class Ejercicio3 {
     public static void main(String[] args) {
         // ? Programa que con un bucle for muestre todos los multiplos de 7 menores que 100
         // ^ Declaramos la variable i (incremento)
-        long i;
+        short i;
 
         // ! Pequeña descripción del programa
         System.out.println("Programa que con un bucle for muestre todos los multiplos de 7 menores que 100");
@@ -12,7 +12,7 @@ public class Ejercicio3 {
         // ? Salto de Línea Triple
         System.out.println("\n" + "-------------------------------" + "\n");
 
-        // ! Bucle for que muestra los números pares desde el 1 al 200
+        // ! Bucle for que muestra los multiplos de 7 menores que 100
         for (i = 7; i < 100; i += 7) {
             System.out.println(i);
         }
